feat(function): 项目TabBar

+++ lib/bean/project_tree_entity.dart
@@ -3,7 +3,7 @@ import 'package:wan_android_getx/generated/json/base/json_convert_content.dart';
 class ProjectTreeEntity with JsonConvert<ProjectTreeEntity> {
 	List<dynamic>? children;
 	double? courseId;
-	double? id;
+	int? id;
 	String? name;
 	double? order;
 	double? parentChapterId;
